requirement.txt
flask==3.0.0
pycryptodome==3.19.0
werkzeug==3.0.1
pyopenssl==23.3.0